refactor: Long -> Int로 변경

+++ app/src/main/java/woowacourse/shopping/data/defaultRepository/DefaultChargeRepository.kt
@@ -7,14 +7,14 @@ import woowacourse.shopping.domain.util.WoowaResult
 import woowacourse.shopping.util.enqueueUtil
 
 class DefaultChargeRepository : ChargeRepository {
-    override fun fetchCharge(callback: (WoowaResult<Long>) -> Unit) {
+    override fun fetchCharge(callback: (WoowaResult<Int>) -> Unit) {
         ServicePool.retrofitService.getCharge().enqueueUtil(
             onSuccess = { callback(WoowaResult.SUCCESS(it.totalCash)) },
             onFailure = { callback(WoowaResult.FAIL(it)) },
         )
     }
 
-    override fun recharge(amount: Long, callback: (WoowaResult<Long>) -> Unit) {
+    override fun recharge(amount: Int, callback: (WoowaResult<Int>) -> Unit) {
         ServicePool.retrofitService.recharge(RequestChargeDto(amount)).enqueueUtil(
             onSuccess = { callback(WoowaResult.SUCCESS(it.totalCash)) },
             onFailure = { callback(WoowaResult.FAIL(it)) },
