+++ app/src/main/java/woowacourse/shopping/data/remote/dto/request/RequestChargeDto.kt
@@ -1,3 +1,3 @@
 package woowacourse.shopping.data.remote.dto.request
 
-data class RequestChargeDto(val cashToCharge: Long)
+data class RequestChargeDto(val cashToCharge: Int)
