+++ app/src/main/java/woowacourse/shopping/data/remote/dto/response/ResponseChargeDto.kt
@@ -1,3 +1,3 @@
 package woowacourse.shopping.data.remote.dto.response
 
-data class ResponseChargeDto(val totalCash: Long)
+data class ResponseChargeDto(val totalCash: Int)
