+++ app/src/main/java/woowacourse/shopping/presentation/ui/myPage/MyPageContract.kt
@@ -2,13 +2,13 @@ package woowacourse.shopping.presentation.ui.myPage
 
 interface MyPageContract {
     interface View {
-        fun showCharge(amount: Long)
+        fun showCharge(amount: Int)
         fun requestRecharge()
         fun showError()
     }
 
     interface Presenter {
         fun fetchCharge()
-        fun recharge(amount: Long)
+        fun recharge(amount: Int)
     }
 }
