+++ domain/src/main/java/woowacourse/shopping/domain/repository/ChargeRepository.kt
@@ -3,6 +3,6 @@ package woowacourse.shopping.domain.repository
 import woowacourse.shopping.domain.util.WoowaResult
 
 interface ChargeRepository {
-    fun fetchCharge(callback: (WoowaResult<Long>) -> Unit)
-    fun recharge(amount: Long, callback: (WoowaResult<Long>) -> Unit)
+    fun fetchCharge(callback: (WoowaResult<Int>) -> Unit)
+    fun recharge(amount: Int, callback: (WoowaResult<Int>) -> Unit)
 }
